pyproject.toml
[project]
name = "dnn-surgery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.0"
dependencies = [
    "grpcio",
    "grpcio-tools",
    "psutil>=7.0.0",
    "pymaxflow>=1.3.2",
    "transformers>=4.54.0",
    "ultralytics>=8.3.177",
]
